index error fix in train function
Proposed are a few changes I implemented to solve issues I ran into. These are not tested on binary data conditions!
- Line 30 of the proposed changes addresses #4
- Line 74 of the proposed changes addresses an undocumented issue I had.

Additionally, the workaround tested by @ropinheiro in #4, shows that user loss_types may be categorical which are not necessary == 'categorical'. A future improvement may be to modify line 41 similar to: 
```if  'categorical' in loss_type:```